+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Added
 
 - log external email addresses when sending
+- allow manually deleting denied checkout requests
 
 ### Fixed